mv final raw data fixtures over to separate module
