+++ wrappers/test/raw_data_fixtures.py
@@ -147,3 +147,27 @@ def annotation_refflat(tmpdir_factory):
     fn = 'annotation/dm6.small.refflat'
     d = tmpdir_for_func(tmpdir_factory)
     return _download_file(fn, d)
+
+
+@pytest.fixture(scope='session')
+def annotation_db(annotation):
+    import gffutils
+    gffutils.create_db(
+        data=annotation, dbfn=annotation + '.db',
+        merge_strategy='merge',
+        id_spec={'transcript': ['transcript_id', 'transcript_symbol'],
+                 'gene': ['gene_id', 'gene_symbol']},
+        gtf_transcript_key='transcript_id',
+        gtf_gene_key='gene_id')
+    return annotation + '.db'
+
+
+@pytest.fixture(scope='session')
+def annotation_bed12(annotation_db):
+    import gffutils
+    db = gffutils.FeatureDB(annotation_db)
+    bed12 = '.'.join(annotation_db.strip().split('.')[:-2]) + '.bed12'
+    with open(bed12, 'w') as handle:
+        for t in db.features_of_type('transcript'):
+            handle.write(db.bed12(t, name_field='transcript_id') + '\n')
+    return bed12
